+++ app/src/main/java/org/simple/clinic/patient/PatientRepository.kt
@@ -644,7 +644,12 @@ class PatientRepository @Inject constructor(
         .findPatientsWithBusinessId(identifier)
         .map { patients -> Optional.ofNullable(patients.firstOrNull()) }
         .toObservable()
+  }
 
+  fun findPatientsWithBusinessId(identifier: String): List<Patient> {
+    return database
+        .patientDao()
+        .findPatientsWithBusinessIdImmediate(identifier)
   }
 
   fun bpPassportForPatient(patientUuid: UUID): Optional<BusinessId> {